chore: extract engine type defs to dedicated crate
