+++ Cargo.toml
@@ -42,6 +42,8 @@ members = [
     "crates/rpc/rpc-testing-util/",
     "crates/rpc/rpc-types/",
     "crates/rpc/rpc-types-compat/",
+    "crates/engine-primitives/",
+    "crates/ethereum-engine-primitives/",
     "crates/node-ethereum/",
     "crates/node-builder/",
     "crates/node-optimism/",
@@ -206,6 +208,8 @@ reth-db = { path = "crates/storage/db" }
 reth-discv4 = { path = "crates/net/discv4" }
 reth-discv5 = { path = "crates/net/discv5" }
 reth-dns-discovery = { path = "crates/net/dns" }
+reth-engine-primitives = { path = "crates/engine-primitives" }
+reth-ethereum-engine-primitives = { path = "crates/ethereum-engine-primitives" }
 reth-node-builder = { path = "crates/node-builder" }
 reth-node-ethereum = { path = "crates/node-ethereum" }
 reth-node-optimism = { path = "crates/node-optimism" }
