+++ crates/engine-primitives/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "reth-node-api"
+name = "reth-engine-primitives"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -12,7 +12,9 @@ workspace = true
 
 [dependencies]
 # reth
-reth-evm.workspace = true
-reth-provider.workspace = true
-reth-db.workspace = true
-reth-engine-primitives.workspace = true
+reth-primitives.workspace = true
+reth-rpc-types.workspace = true
+
+# misc
+serde.workspace = true
+thiserror.workspace = true
