+++ crates/engine-primitives/src/lib.rs
@@ -1,4 +1,12 @@
-//! This contains the [EngineTypes] trait and implementations for ethereum mainnet types.
+//! Traits, validation methods, and helper types used to abstract over engine types.
+
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
+    html_favicon_url = "https://avatars0.githubusercontent.com/u/97369466?s=256",
+    issue_tracker_base_url = "https://github.com/paradigmxyz/reth/issues/"
+)]
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 use core::fmt;
 use reth_primitives::ChainSpec;
