+++ crates/ethereum-engine-primitives/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "reth-ethereum-engine-primitives"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# reth
+reth-primitives.workspace = true
+reth-engine-primitives.workspace = true
+reth-rpc-types.workspace = true
+reth-rpc-types-compat.workspace = true
+revm-primitives.workspace = true
+alloy-rlp.workspace = true
+
+# misc
+serde.workspace = true
+sha2.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
