+++ crates/ethereum-engine-primitives/src/lib.rs
@@ -1,8 +1,20 @@
-use reth_node_api::{
+//! Ethereum specifc
+
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
+    html_favicon_url = "https://avatars0.githubusercontent.com/u/97369466?s=256",
+    issue_tracker_base_url = "https://github.com/paradigmxyz/reth/issues/"
+)]
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+
+mod payload;
+pub use payload::{EthBuiltPayload, EthPayloadBuilderAttributes};
+
+use reth_engine_primitives::{
     validate_version_specific_fields, EngineApiMessageVersion, EngineObjectValidationError,
     EngineTypes, PayloadOrAttributes,
 };
-use reth_payload_builder::{EthBuiltPayload, EthPayloadBuilderAttributes};
 use reth_primitives::ChainSpec;
 use reth_rpc_types::{
     engine::{
