+++ crates/node-ethereum/Cargo.toml
@@ -12,12 +12,10 @@ workspace = true
 
 [dependencies]
 # reth
-reth-primitives.workspace = true
 reth-payload-builder.workspace = true
+reth-ethereum-engine-primitives.workspace = true
 reth-basic-payload-builder.workspace = true
 reth-ethereum-payload-builder.workspace = true
-reth-rpc-types.workspace = true
-reth-node-api.workspace = true
 reth-node-builder.workspace = true
 reth-tracing.workspace = true
 reth-provider.workspace = true
@@ -27,7 +25,6 @@ reth-evm-ethereum.workspace = true
 
 # misc
 eyre.workspace = true
-serde.workspace = true
 
 [dev-dependencies]
 reth-db.workspace = true
