+++ crates/payload/builder/Cargo.toml
@@ -17,13 +17,9 @@ reth-primitives.workspace = true
 reth-rpc-types.workspace = true
 reth-transaction-pool.workspace = true
 reth-interfaces.workspace = true
-reth-rpc-types-compat.workspace = true
 reth-provider.workspace = true
-reth-node-api.workspace = true
-
-# ethereum
-alloy-rlp.workspace = true
-revm-primitives.workspace = true
+reth-engine-primitives.workspace = true
+reth-ethereum-engine-primitives.workspace = true
 
 # async
 tokio = { workspace = true, features = ["sync"] }
@@ -36,7 +32,6 @@ metrics.workspace = true
 
 # misc
 thiserror.workspace = true
-sha2.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
