+++ crates/payload/builder/src/events.rs
@@ -1,5 +1,5 @@
 use futures_util::Stream;
-use reth_node_api::EngineTypes;
+use reth_engine_primitives::EngineTypes;
 use tokio::sync::broadcast;
 use tokio_stream::{
     wrappers::{errors::BroadcastStreamRecvError, BroadcastStream},
