+++ crates/payload/builder/src/traits.rs
@@ -1,7 +1,7 @@
 //! Trait abstractions used by the payload crate.
 
 use crate::error::PayloadBuilderError;
-use reth_node_api::{BuiltPayload, PayloadBuilderAttributes};
+use reth_engine_primitives::{BuiltPayload, PayloadBuilderAttributes};
 use reth_provider::CanonStateNotification;
 use std::future::Future;
